+++ packages/sources/stader-address-list/src/config/index.ts
@@ -1,13 +1,28 @@
-import { Requester, util } from '@chainlink/ea-bootstrap'
-import type { Config } from '@chainlink/ea-bootstrap'
+import { AdapterConfig } from '@chainlink/external-adapter-framework/config'
 
-export const NAME = 'STADER_ADDRESS_LIST'
-
-export const DEFAULT_ENDPOINT = 'address'
-
-export const makeConfig = (prefix?: string): Config => {
-  const config = Requester.getDefaultConfig(prefix)
-  config.defaultEndpoint = DEFAULT_ENDPOINT
-  config.rpcUrl = util.getRequiredEnv('RPC_URL')
-  return config
-}
+export const config = new AdapterConfig(
+  {
+    RPC_URL: {
+      type: 'string',
+      description:
+        'The RPC URL to connect to the EVM chain the address manager contract is deployed to.',
+      required: true,
+    },
+    CHAIN_ID: {
+      type: 'number',
+      description: 'The chain id to connect to',
+      default: 1,
+    },
+    BACKGROUND_EXECUTE_MS: {
+      description:
+        'The amount of time the background execute should sleep before performing the next request',
+      type: 'number',
+      default: 10_000,
+    },
+  },
+  {
+    envDefaultOverrides: {
+      BACKGROUND_EXECUTE_TIMEOUT: 180_000,
+    },
+  },
+)