+++ packages/composites/defi-pulse/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/defi-pulse-adapter
 
+## 1.2.49
+
+### Patch Changes
+
+- Updated dependencies [77ad946a9]
+  - @chainlink/token-allocation-adapter@1.10.6
+
 ## 1.2.48
 
 ### Patch Changes