+++ packages/sources/icap/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/icap-adapter
 
+## 1.1.1
+
+### Patch Changes
+
+- 77ad946a9: Bumped framework version
+- Updated dependencies [77ad946a9]
+  - @chainlink/tp-adapter@1.1.1
+
 ## 1.1.0
 
 ### Minor Changes